Change some paramters in launch RPC from uint16_t to uint32_t for consistency with job allocate.

+++ branches/mc.struct/src/common/slurm_protocol_defs.h
@@ -371,7 +371,7 @@ typedef struct last_update_msg {
 typedef struct launch_tasks_request_msg {
 	uint32_t  job_id;
 	uint32_t  job_step_id;
-	uint32_t  nnodes;	/* number of nodes in this job step       */
+	uint16_t  nnodes;	/* number of nodes in this job step       */
 	uint32_t  nprocs;	/* number of processes in this job step   */
 	uint32_t  uid;
 	uint32_t  gid;
@@ -379,14 +379,14 @@ typedef struct launch_tasks_request_msg {
 	uint16_t  envc;
 	uint16_t  argc;
 	uint16_t  multi_prog;
-	uint32_t  *cpus_allocated;
-	uint32_t  max_sockets;
-	uint32_t  max_cores;
-	uint32_t  max_threads;
-	uint32_t  cpus_per_task;
-	uint32_t  ntasks_per_node;
-	uint32_t  ntasks_per_socket;
-	uint32_t  ntasks_per_core;
+	uint16_t  *cpus_allocated;
+	uint16_t  max_sockets;
+	uint16_t  max_cores;
+	uint16_t  max_threads;
+	uint16_t  cpus_per_task;
+	uint16_t  ntasks_per_node;
+	uint16_t  ntasks_per_socket;
+	uint16_t  ntasks_per_core;
 	char    **env;
 	char    **argv;
 	char     *cwd;
@@ -399,7 +399,7 @@ typedef struct launch_tasks_request_msg {
 
         /* Distribution at the lowest level of logical processor (lllp) */
 	uint16_t task_dist;  /* --distribution=, -m dist	*/
-	uint32_t plane_size; /* lllp distribution -> plane_size for
+	uint16_t plane_size; /* lllp distribution -> plane_size for
 			      * when -m plane=<# of lllp per plane> */      
 	uint16_t  task_flags;
 	uint32_t **global_task_ids;
@@ -421,7 +421,7 @@ typedef struct launch_tasks_request_msg {
 	char     *task_prolog;
 	char     *task_epilog;
 
-	uint32_t   slurmd_debug; /* remote slurmd debug level */
+	uint16_t   slurmd_debug; /* remote slurmd debug level */
 
 	slurm_cred_t cred;	/* job credential            */
 	switch_jobinfo_t switch_job;	/* switch credential for the job */
